:recycle: (Storybook) Remove ButtonTertiary to use black or primary
