+++ src/features/cheatcodes/pages/AppComponents/AppComponents.tsx
@@ -30,7 +30,7 @@ import { ButtonPrimaryWhite } from 'ui/components/buttons/ButtonPrimaryWhite'
 import { ButtonQuaternary } from 'ui/components/buttons/ButtonQuaternary'
 import { ButtonQuaternaryBlack } from 'ui/components/buttons/ButtonQuaternaryBlack'
 import { ButtonSecondary } from 'ui/components/buttons/ButtonSecondary'
-import { ButtonTertiary } from 'ui/components/buttons/ButtonTertiary'
+import { ButtonTertiaryPrimary } from 'ui/components/buttons/ButtonTertiaryPrimary'
 import { ButtonTertiaryWhite } from 'ui/components/buttons/ButtonTertiaryWhite'
 import { ButtonWithLinearGradient } from 'ui/components/buttons/ButtonWithLinearGradient'
 import { ExternalLink } from 'ui/components/buttons/externalLink/ExternalLink'
@@ -180,14 +180,14 @@ export const AppComponents: FunctionComponent = () => {
           {/* Buttons: Tertiary */}
           <Typo.Title4>Button - Theme Tertiary</Typo.Title4>
           <Spacer.Column numberOfSpaces={1} />
-          <ButtonTertiary
+          <ButtonTertiaryPrimary
             wording="Triggers all active buttons"
             icon={Close}
             isLoading={buttonIsLoading}
             onPress={onTriggerFakeLoading}
           />
           <Spacer.Column numberOfSpaces={1} />
-          <ButtonTertiary wording="Disabled" onPress={onButtonPress} icon={Close} disabled />
+          <ButtonTertiaryPrimary wording="Disabled" onPress={onButtonPress} icon={Close} disabled />
           <Spacer.Column numberOfSpaces={1} />
           <ButtonTertiaryWhite
             wording="White tertiary button"
