+++ src/ui/components/buttons/ButtonTertiary.stories.tsx
@@ -11,24 +11,19 @@ import { StoryContainer } from 'ui/storybook/StoryContainer'
 import { EditPen } from 'ui/svg/icons/EditPen'
 import { Email } from 'ui/svg/icons/Email'
 
-import { ButtonTertiary } from './ButtonTertiary'
-
 export default {
   title: 'ui/buttons/ButtonTertiary',
-  component: ButtonTertiary,
+  component: ButtonTertiaryPrimary,
   argTypes: {
     icon: selectArgTypeFromObject({
       Email,
       EditPen,
     }),
   },
-} as ComponentMeta<typeof ButtonTertiary>
+} as ComponentMeta<typeof ButtonTertiaryPrimary>
 
-const Template: ComponentStory<typeof ButtonTertiary> = (props) => (
+const Template: ComponentStory<typeof ButtonTertiaryPrimary> = (props) => (
   <React.Fragment>
-    <StoryContainer title="ButtonTertiary">
-      <ButtonTertiary {...props} />
-    </StoryContainer>
     <StoryContainer title="ButtonTertiaryBlack">
       <ButtonTertiaryBlack {...props} />
     </StoryContainer>
@@ -54,7 +49,7 @@ Default.args = {
 Default.parameters = {
   docs: {
     source: {
-      code: '<ButtonTertiary wording="Confirmer" />',
+      code: '<ButtonTertiaryPrimary wording="Confirmer" />',
     },
   },
 }
@@ -67,7 +62,7 @@ Loading.args = {
 Loading.parameters = {
   docs: {
     source: {
-      code: '<ButtonTertiary wording="Confirmer" isLoading />',
+      code: '<ButtonTertiaryPrimary wording="Confirmer" isLoading />',
     },
   },
 }
@@ -80,7 +75,7 @@ Tall.args = {
 Tall.parameters = {
   docs: {
     source: {
-      code: '<ButtonTertiary wording="Confirmer" buttonHeight="tall" />',
+      code: '<ButtonTertiaryPrimary wording="Confirmer" buttonHeight="tall" />',
     },
   },
 }
@@ -93,7 +88,7 @@ WithIcon.args = {
 WithIcon.parameters = {
   docs: {
     source: {
-      code: '<ButtonTertiary wording="Confirmer" icon={Email} />',
+      code: '<ButtonTertiaryPrimary wording="Confirmer" icon={Email} />',
     },
   },
 }
@@ -107,7 +102,7 @@ DisabledWithIcon.args = {
 DisabledWithIcon.parameters = {
   docs: {
     source: {
-      code: '<ButtonTertiary wording="Confirmer" disabled={true} icon={Email} />',
+      code: '<ButtonTertiaryPrimary wording="Confirmer" disabled={true} icon={Email} />',
     },
   },
 }
